+++ core/server/master/src/test/java/alluxio/master/metastore/caching/CachingInodeStoreMockedBackingStoreTest.java
@@ -56,7 +56,7 @@ import java.util.concurrent.ThreadLocalRandom;
 import java.util.concurrent.atomic.AtomicInteger;
 
 public class CachingInodeStoreMockedBackingStoreTest {
-  private static final long CACHE_SIZE = 20;
+  private static final int CACHE_SIZE = 20;
   private static final long TEST_INODE_ID = 5;
   private static final MutableInodeDirectory TEST_INODE_DIR =
       MutableInodeDirectory.create(TEST_INODE_ID, 0, "name", CreateDirectoryContext.defaults());
